# web-app/public/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API Documentation",
    "version": "1.0.0",
    "description": "This is the OpenAPI specification for your project."
  },
  "servers": [
    {
      "url": "https://api.sokosumi.com/api",
      "description": "Production server"
    },
    {
      "url": "http://localhost:3000/api",
      "description": "Local development server"
    }
  ],
  "components": {
...
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "description": "API key required in the x-api-key header"
      }
    },
...
    "responses": {
      "400": {
        "description": "Bad Request",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Invalid request parameters"
                },
                "code": {
                  "type": "string",
                  "example": "400"
                }
              }
            }
          }
        }
      },
...
      "500": {
        "description": "Internal Server Error",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "An unexpected error occurred"
                },
                "code": {
                  "type": "string",
                  "example": "500"
                }
              }
            }
          }
        }
      }
    }
  },
  "paths": {
...
    "/v1/jobs": {
      "get": {
        "operationId": "get-v1-jobs",
        "summary": "List jobs",
        "description": "Retrieves all jobs that belong to the authenticated user. Supports optional query parameters for filtering:",
        "tags": [
          "Jobs"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "agentId",
            "schema": {
              "type": "string"
            },
            "required": false
          },
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string"
            },
            "required": false
          }
        ],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "createdAt": {
                            "type": "string",
                            "description": "ISO date"
                          },
                          "updatedAt": {
                            "type": "string",
                            "description": "ISO date"
                          },
                          "name": {
                            "type": "string",
                            "nullable": true
                          },
                          "status": {
                            "type": "string",
                            "description": "JobStatus"
                          },
                          "agentId": {
                            "type": "string"
                          },
                          "userId": {
                            "type": "string"
                          },
                          "organizationId": {
                            "type": "string",
                            "nullable": true
                          },
                          "agentJobId": {
                            "type": "string"
                          },
                          "agentJobStatus": {
                            "type": "string",
                            "nullable": true
                          },
                          "onChainStatus": {
                            "type": "string",
                            "nullable": true
                          },
                          "input": {
                            "type": "string"
                          },
                          "output": {
                            "type": "string",
                            "nullable": true
                          },
                          "startedAt": {
                            "type": "string",
                            "description": "ISO date"
                          },
                          "completedAt": {
                            "type": "string",
                            "nullable": true,
                            "description": "ISO date"
                          },
                          "resultSubmittedAt": {
                            "type": "string",
                            "nullable": true,
                            "description": "ISO date"
                          },
                          "isDemo": {
                            "type": "boolean"
                          },
                          "price": {
                            "type": "object",
                            "properties": {
                              "credits": {
                                "type": "number"
                              },
                              "includedFee": {
                                "type": "number"
                              }
                            },
                            "nullable": true
                          },
                          "refund": {
                            "type": "object",
                            "properties": {
                              "credits": {
                                "type": "number"
                              },
                              "includedFee": {
                                "type": "number"
                              }
                            },
                            "nullable": true
                          },
                          "jobStatusSettled": {
                            "type": "boolean"
                          }
                        }
                      }
                    },
                    "timestamp": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
...
  }
}